Split code between dags and operator implementation packages
This reorganizes dataflow code into operators source code (either
python functions or SQL queries in our case) and DAGs that configure
and combine operators.

Previously separate `meta` classes that defined configuration for each
DAG have been moved to the `dags` folder. The now subclass common base
classes which have a class method generating the DAG itself. This should
make their goal clear and connects them to the resulting DAG. This
also simplifies pipeline discovery since it allows us to iterate over
base class subclasses instead of looking for keywords in module attributes.

DAGs are otherwise registered the same way as before - by adding a key
to the globals dict in each dag file.

Hopefully this makes the overall layout a bit clearer since it follows
airflow concepts of DAGs and operators.
