+++ dataflow/dags/dataset_pipelines.py
@@ -1,371 +1,46 @@
 """A module that defines Airflow DAGS for dataset pipelines."""
 
-import json
-import logging
-import time
-
-from datetime import timedelta
+from datetime import datetime, timedelta
 
 from airflow import DAG
-from airflow.hooks.postgres_hook import PostgresHook
-from airflow.models import Variable
 from airflow.operators.python_operator import PythonOperator
-from jinja2 import Template
-from mohawk import Sender
-from mohawk.exc import HawkFail
-from psycopg2 import sql
-import redis
-import requests
 
 from dataflow import constants
-from dataflow.meta import dataset_pipelines
-from dataflow.utils import get_defined_pipeline_classes_by_key
-
-
-dataset_pipeline_classes = get_defined_pipeline_classes_by_key(
-    dataset_pipelines, 'DatasetPipeline'
+from dataflow.operators.dataset import (
+    run_fetch,
+    create_tables,
+    execute_insert_into,
+    insert_from_temporary_table,
 )
 
-credentials = {
-    'id': constants.HAWK_ID,
-    'key': constants.HAWK_KEY,
-    'algorithm': constants.HAWK_ALGORITHM,
-}
 
+class BaseDatasetPipeline:
+    @classmethod
+    def get_dag(pipeline):
+        run_fetch_task_id = f'RunFetch{pipeline.__name__}'
 
-def get_redis_client():
-    """Returns redis client from connection URL"""
-    return redis.from_url(url=constants.REDIS_URL)
-
-
-def mark_task_failed(task_instance):
-    """Marks task as failed and delete variables set by the task"""
-    redis_client = get_redis_client()
-    task_instance.xcom_push(key='state', value=False)
-    for i in range(0, redis_client.llen(run_fetch_task_id)):
-        Variable.delete(redis_client.lindex(run_fetch_task_id, i))
-    redis_client.delete(run_fetch_task_id)
-
-
-def run_fetch(source_url, run_fetch_task_id=None, task_instance=None, **kwargs):
-    """Fetch data from source.
-
-    Args:
-        source_url (str): URL for API Endpoint to fetch data from source.
-
-    Source endpoint has to accept GET request and respond with HTTP 200 OK for success.
-    Needs to be paginated, response is expected to have below structure;
-    {
-        'next': <link_to_next_page>,
-        'results': [list of dict]
-    }
-
-    Notes:
-    XCOM isn't used to transfer data between tasks because it is not built to handle
-    very large data transfer between tasks.
-    Saving fetched records into a file would prevent us from scaling with celery, saving into
-    a single variable causes worker shutdown due to high memory usage. That's why,
-    this saves paginated response into indexed named variables and returns variable names
-    to be consumed by the task which inserts data into regarding table. Alternatives are;
-    - Shared network storage
-    - S3 (Security risk)
-
-    Example source_url.
-        source_url = https://datahub-api-demo.london.cloudapps.digital/v4/datasets/omis-dataset
-    TODO:
-        By the impletation of other Datasets pipeline, there will be more generic structure to
-        support various pipeline types.
-
-    """
-    redis_client = get_redis_client()
-    # Clear any leftover requests from previous task runs
-    redis_client.delete(run_fetch_task_id)
-
-    index = 0
-    while True:
-        sender = Sender(credentials, source_url, 'get', always_hash_content=False)
-
-        logging.info(f'Fetching page {source_url}')
-        response = requests.get(
-            source_url, headers={'Authorization': sender.request_header}
+        dag = DAG(
+            pipeline.__name__,
+            catchup=False,
+            default_args={
+                'owner': 'airflow',
+                'depends_on_past': False,
+                'email_on_failure': False,
+                'email_on_retry': False,
+                'retries': 0,
+                'retry_delay': timedelta(minutes=5),
+            },
+            start_date=pipeline.start_date,
+            end_date=pipeline.end_date,
+            schedule_interval=pipeline.schedule_interval,
+            max_active_runs=1,
+            user_defined_macros={
+                'table_name': pipeline.table_name,
+                'field_mapping': pipeline.field_mapping,
+                'run_fetch_task_id': run_fetch_task_id,
+            },
         )
-        if response.status_code != 200:
-            mark_task_failed(task_instance)
-            raise Exception(
-                f'GET request to {source_url} is unsuccessful\n'
-                f'Message: {response.text}'
-            )
-        try:
-            sender.accept_response(
-                response.headers['Server-Authorization'],
-                content=response.content,
-                content_type=response.headers['Content-Type'],
-            )
-        except HawkFail as e:
-            mark_task_failed(task_instance)
-            raise Exception(f'HAWK Authentication failed {str(e)}')
 
-        response_json = response.json()
-        if 'results' not in response_json or 'next' not in response_json:
-            mark_task_failed(task_instance)
-            raise Exception('Unexpected response structure')
-
-        key = f'{run_fetch_task_id}{index}'
-        Variable.set(key, response_json['results'], serialize_json=True)
-        redis_client.rpush(run_fetch_task_id, key)
-        next_page = response_json['next']
-        if next_page:
-            index += 1
-            source_url = next_page
-        else:
-            break
-
-    logging.info('Fetching from source completed')
-    task_instance.xcom_push(key='state', value=True)
-
-
-def create_tables(target_db, table_name=None, field_mapping=None, **kwargs):
-    """
-    Create a temporary table to be copied over to the target table `table_name`.
-    """
-    temp_table_name = F'{table_name}_tmp'
-    create_table_sql = """
-        CREATE TABLE IF NOT EXISTS {{ table_name }} (
-        {% for _, tt_field_name, tt_field_constraints in field_mapping %}
-            {{ tt_field_name }} {{ tt_field_constraints }}{{ "," if not loop.last }}
-        {% endfor %}
-        );
-    """
-
-    try:
-        target_db_conn = PostgresHook(postgres_conn_id=target_db).get_conn()
-        target_db_cursor = target_db_conn.cursor()
-        logging.info(f'Creating temporary table {temp_table_name}')
-        target_db_cursor.execute(
-            sql.SQL('DROP TABLE IF EXISTS {}').format(sql.Identifier(temp_table_name))
-        )
-        target_db_cursor.execute(
-            Template(create_table_sql).render(
-                table_name=sql.Identifier(temp_table_name).as_string(target_db_conn),
-                field_mapping=field_mapping,
-            )
-        )
-        logging.info(f'Creating target table {table_name}')
-        target_db_cursor.execute(
-            Template(create_table_sql).render(
-                table_name=sql.Identifier(table_name).as_string(target_db_conn),
-                field_mapping=field_mapping,
-            )
-        )
-        target_db_conn.commit()
-
-    # TODO: Gotta Catch'm all
-    except Exception as e:
-        logging.error(f'Exception: {e}')
-        target_db_conn.rollback()
-        raise
-
-    finally:
-        if target_db_conn:
-            target_db_cursor.close()
-            target_db_conn.close()
-
-
-def insert_from_temporary_table(
-    target_db, table_name=None, task_instance=None, run_fetch_task_id=None, **kwargs
-):
-    """
-    Insert data from temporary table into `table_name`.
-    The rational behind is not doing any modification on target table
-    before we make sure fetching from source and insertion is successful.
-    """
-    temp_table_name = F'{table_name}_tmp'
-    rename_table_sql = """
-        TRUNCATE TABLE {table_name};
-        INSERT INTO {table_name}
-        SELECT * FROM {temp_table_name};
-        DROP TABLE {temp_table_name};
-    """
-    fetcher_state = task_instance.xcom_pull(key='state', task_ids=run_fetch_task_id)
-    inserter_state = True
-    # Check all insertion tasks are completed successfully.
-    for index in range(constants.INGEST_TASK_CONCURRENCY):
-        inserter_state = inserter_state and task_instance.xcom_pull(
-            key='state', task_ids=f'execute-insert-into-{index}'
-        )
-    if fetcher_state is True and inserter_state is True:
-        logging.info(f'Inserting data from {temp_table_name} into {table_name}')
-        try:
-            target_db_conn = PostgresHook(postgres_conn_id=target_db).get_conn()
-            target_db_cursor = target_db_conn.cursor()
-            target_db_cursor.execute(
-                rename_table_sql.format(
-                    table_name=sql.Identifier(table_name).as_string(target_db_conn),
-                    temp_table_name=sql.Identifier(temp_table_name).as_string(
-                        target_db_conn
-                    ),
-                )
-            )
-            target_db_conn.commit()
-
-        # TODO: Gotta Catch'm all
-        except Exception as e:
-            logging.error(f'Exception: {e}')
-            mark_task_failed(task_instance)
-            target_db_conn.rollback()
-            raise
-
-        finally:
-            if target_db_conn:
-                target_db_cursor.close()
-                target_db_conn.close()
-    else:
-        logging.info('Fetcher or inserter failed!')
-        mark_task_failed(task_instance)
-
-
-def execute_insert_into(
-    target_db,
-    table_name=None,
-    run_fetch_task_id=None,
-    field_mapping=None,
-    task_instance=None,
-    **kwargs,
-):
-    """Inserts each paginated response data into target database table.
-    Polls to find variable hasn't been processed, generates regarding sql statement to
-    insert data in, incrementally waits for new variables.
-    Success depends on fetcher task completion.
-    """
-    temp_table_name = F'{table_name}_tmp'
-    insert_into_sql = """
-        INSERT INTO {{ temp_table_name }} (
-        {% for _, tt_field_name, _ in field_mapping %}
-            {{ tt_field_name }}{{ "," if not loop.last }}
-        {% endfor %}
-        )
-        VALUES
-        {% for record in record_subset %}
-        (
-            {% for st_field_name, _, _ in field_mapping %}
-                {% if not record[st_field_name] or record[st_field_name] == 'None' %}
-                    NULL
-                {% else %}
-                    {{ record[st_field_name] }}
-                {% endif %}
-                {{ "," if not loop.last }}
-            {% endfor %}
-        {{ ")," if not loop.last }}
-        {% endfor %}
-        );
-    """
-    # Give some initial time to fetch task to get a page and save it into variable
-    time.sleep(3)
-    # Used for providing incremental wait
-    sleep_time = 5
-    number_of_runs = 1
-    redis_client = get_redis_client()
-    try:
-        target_db_conn = PostgresHook(postgres_conn_id=target_db).get_conn()
-        target_db_cursor = target_db_conn.cursor()
-        while True:
-            var_name = redis_client.lpop(run_fetch_task_id)
-            if var_name:
-                logging.info(f'Processing page {var_name}')
-                sleep_time = 5
-                var_name = var_name.decode('utf-8')
-                try:
-                    record_subset = json.loads(Variable.get(var_name))
-                except KeyError:
-                    logging.info(f'Page {var_name} does not exist. Moving on.')
-                    continue
-
-                escaped_record_subset = []
-                for record in record_subset:
-                    escaped_record = {}
-                    for key, value in record.items():
-                        if value and value != 'None':
-                            escaped_record[key] = sql.Literal(value).as_string(
-                                target_db_conn
-                            )
-                        else:
-                            escaped_record[key] = sql.Literal(None).as_string(
-                                target_db_conn
-                            )
-                    escaped_record_subset.append(escaped_record)
-
-                exec_sql = Template(insert_into_sql).render(
-                    temp_table_name=sql.Identifier(temp_table_name).as_string(
-                        target_db_conn
-                    ),
-                    field_mapping=field_mapping,
-                    record_subset=escaped_record_subset,
-                )
-                target_db_cursor.execute(exec_sql)
-                logging.info(f'Page {var_name} ingested successfully')
-                Variable.delete(var_name)
-            else:
-                # Check if fetch task completed successfully, if it's, break out of loop and commit
-                # the transaction because there is no more page to process.
-                # If it's failed raise Exception so that transaction will be rollbacked
-                state = task_instance.xcom_pull(key='state', task_ids=run_fetch_task_id)
-                logging.info(f'Fetcher task {run_fetch_task_id} state is "{state}"')
-                if state is False:
-                    mark_task_failed(task_instance)
-                    raise Exception(f'Fetcher task {run_fetch_task_id} failed!')
-                elif state is True:
-                    logging.info(f'Fetcher task {run_fetch_task_id} has completed.')
-                    break
-                else:
-                    logging.info(
-                        f'Sleeping for {sleep_time} so task {run_fetch_task_id} can catchup'
-                    )
-                    sleep_time = sleep_time * number_of_runs
-                    time.sleep(sleep_time)
-                    number_of_runs += 1
-
-        target_db_conn.commit()
-        task_instance.xcom_push(key='state', value=True)
-
-    # TODO: Gotta Catch'm all
-    except Exception as e:
-        logging.error(f'Exception: {e}')
-        mark_task_failed(task_instance)
-        target_db_conn.rollback()
-        raise
-
-    finally:
-        if target_db_conn:
-            target_db_cursor.close()
-            target_db_conn.close()
-
-
-default_args = {
-    'owner': 'airflow',
-    'depends_on_past': False,
-    'email_on_failure': False,
-    'email_on_retry': False,
-    'retries': 0,
-    'retry_delay': timedelta(minutes=5),
-}
-
-for pipeline in dataset_pipeline_classes:
-    run_fetch_task_id = f'RunFetch{pipeline.__name__}'
-
-    with DAG(
-        pipeline.__name__,
-        catchup=False,
-        default_args=default_args,
-        start_date=pipeline.start_date,
-        end_date=pipeline.end_date,
-        schedule_interval=pipeline.schedule_interval,
-        max_active_runs=1,
-        user_defined_macros={
-            'table_name': pipeline.table_name,
-            'field_mapping': pipeline.field_mapping,
-            'run_fetch_task_id': run_fetch_task_id,
-        },
-    ) as dag:
         t1 = PythonOperator(
             task_id=run_fetch_task_id,
             python_callable=run_fetch,
@@ -398,5 +73,335 @@ for pipeline in dataset_pipeline_classes:
             op_args=[f'{pipeline.target_db}'],
         )
 
-        tend << insert_group << t3
-        globals()[pipeline.__name__] = dag
+        dag << t1
+        dag << tend << insert_group << t3
+
+        return dag
+
+
+class OMISDatasetPipeline(BaseDatasetPipeline):
+    """Pipeline meta object for OMISDataset."""
+
+    table_name = 'omis_dataset'
+    source_url = '{0}/v4/dataset/omis-dataset'.format(constants.DATAHUB_BASE_URL)
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('cancellation_reason__name', 'cancellation_reason', 'text'),
+        ('cancelled_on', 'cancelled_date', 'timestamp with time zone'),
+        ('company_id', 'company_id', 'uuid'),
+        ('completed_on', 'completion_date', 'timestamp with time zone'),
+        ('contact_id', 'contact_id', 'uuid'),
+        ('created_by__dit_team_id', 'dit_team_id', 'uuid'),
+        ('created_on', 'created_date', 'timestamp with time zone'),
+        ('delivery_date', 'delivery_date', 'date'),
+        ('id', 'id', 'uuid primary key'),
+        ('invoice__subtotal_cost', 'subtotal', 'integer'),
+        ('paid_on', 'payment_received_date', 'timestamp with time zone'),
+        ('primary_market__name', 'market', 'text'),
+        ('reference', 'omis_order_reference', 'character varying(100)'),
+        ('sector_name', 'sector', 'character varying(255)'),
+        ('services', 'services', 'text'),
+        ('status', 'order_status', 'character varying(100)'),
+        ('subtotal_cost', 'net_price', 'integer'),
+        ('uk_region__name', 'uk_region', 'text'),
+    ]
+
+
+class InvestmentProjectsDatasetPipeline(BaseDatasetPipeline):
+    """Pipeline meta object for InvestmentProjectsDataset."""
+
+    table_name = 'investment_projects_dataset'
+    source_url = '{0}/v4/dataset/investment-projects-dataset'.format(
+        constants.DATAHUB_BASE_URL
+    )
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('actual_land_date', 'actual_land_date', 'date'),
+        ('actual_uk_region_names', 'actual_uk_regions', 'text'),
+        ('allow_blank_possible_uk_regions', 'possible_uk_regions', 'boolean'),
+        ('anonymous_description', 'anonymous_description', 'text'),
+        (
+            'associated_non_fdi_r_and_d_project__id',
+            'associated_non_fdi_r_and_d_project_id',
+            'character varying(100)',
+        ),
+        ('average_salary__name', 'average_salary', 'text'),
+        ('business_activity_names', 'business_activities', 'text'),
+        ('client_relationship_manager_id', 'client_relationship_manager_id', 'uuid'),
+        ('clint_requirements', 'client_requirements', 'text'),
+        ('competing_countries', 'competing_countries', 'text'),
+        ('created_by_id', 'created_by_id', 'uuid'),
+        ('created_on', 'created_on', 'timestamp with time zone'),
+        ('delivery_partner_names', 'delivery_partners', 'text'),
+        ('description', 'description', 'text'),
+        ('estimated_land_date', 'estimated_land_date', 'date'),
+        ('export_revenue', 'export_revenue', 'boolean'),
+        ('fdi_type__name', 'fdi_type', 'text'),
+        ('fdi_value__name', 'fdi_value', 'text'),
+        ('foreign_equity_investment', 'foreign_equity_investment', 'decimal'),
+        ('government_assistance', 'government_assistance', 'boolean'),
+        ('gross_value_added', 'gross_value_added', 'decimal'),
+        ('gva_multiplier__multiplier', 'gva_multiplier', 'decimal'),
+        ('id', 'id', 'uuid primary key'),
+        ('investment_type__name', 'investment_type', 'text'),
+        ('investor_company_id', 'investor_company_id', 'uuid'),
+        (
+            'investor_company_sector',
+            'investor_company_sector',
+            'character varying(255)',
+        ),
+        ('investor_type__name', 'investor_type', 'text'),
+        ('level_of_involvement_name', 'level_of_involvement', 'text'),
+        ('likelihood_to_land__name', 'likelihood_to_land', 'text'),
+        ('modified_by_id', 'modified_by_id', 'uuid'),
+        ('modified_on', 'modified_on', 'timestamp with time zone'),
+        ('name', 'name', 'character varying(255) NOT NULL'),
+        ('new_tech_to_uk', 'new_tech_to_uk', 'boolean'),
+        ('non_fdi_r_and_d_budget', 'non_fdi_r_and_d_budget', 'boolean'),
+        ('number_new_jobs', 'number_new_jobs', 'integer'),
+        ('number_safeguarded_jobs', 'number_safeguarded_jobs', 'integer'),
+        ('project_arrived_in_triage_on', 'project_arrived_in_triage_on', 'date'),
+        ('project_assurance_adviser_id', 'project_assurance_adviser_id', 'uuid'),
+        ('project_manager_id', 'project_manager_id', 'uuid'),
+        ('project_reference', 'project_reference', 'text'),
+        ('proposal_deadline', 'proposal_deadline', 'date'),
+        ('r_and_d_budget', 'r_and_d_budget', 'boolean'),
+        ('referral_source_activity__name', 'referral_source_activity', 'text'),
+        (
+            'referral_source_activity_marketing__name',
+            'referral_source_activity_marketing',
+            'text',
+        ),
+        (
+            'referral_source_activity_website__name',
+            'referral_source_activity_website',
+            'text',
+        ),
+        ('sector_name', 'sector', 'character varying(255)'),
+        ('specific_programme__name', 'specific_programme', 'text'),
+        ('stage__name', 'stage', 'text NOT NULL'),
+        ('status', 'status', 'character varying(255)'),
+        ('strategic_driver_names', 'strategic_drivers', 'text'),
+        ('team_member_ids', 'team_member_ids', 'text []'),
+        ('total_investment', 'total_investment', 'decimal'),
+        ('uk_company_id', 'uk_company_id', 'uuid'),
+        ('uk_company_sector', 'uk_company_sector', 'character varying(255)'),
+    ]
+
+
+class InteractionsDatasetPipeline(BaseDatasetPipeline):
+    table_name = 'interactions_dataset'
+    source_url = '{}/v4/dataset/interactions-dataset'.format(constants.DATAHUB_BASE_URL)
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('adviser_ids', 'adviser_ids', 'text []'),
+        (
+            'communication_channel__name',
+            'communication_channel',
+            'character varying(255)',
+        ),
+        ('company_id', 'company_id', 'uuid'),
+        ('contact_ids', 'contact_ids', 'text []'),
+        ('created_on', 'created_on', 'timestamp with time zone'),
+        ('date', 'interaction_date', 'date'),
+        ('event_id', 'event_id', 'uuid'),
+        ('grant_amount_offered', 'grant_amount_offered', 'decimal'),
+        ('id', 'id', 'uuid primary key'),
+        ('interaction_link', 'interaction_link', 'character varying(255)'),
+        ('investment_project_id', 'investment_project_id', 'uuid'),
+        ('kind', 'interaction_kind', 'character varying(255)'),
+        ('net_company_receipt', 'net_company_receipt', 'decimal'),
+        ('notes', 'interaction_notes', 'text'),
+        ('sector', 'sector', 'character varying(255)'),
+        (
+            'service_delivery_status__name',
+            'service_delivery_status',
+            'character varying(255)',
+        ),
+        ('service_delivery', 'service_delivery', 'character varying(255)'),
+        ('subject', 'interaction_subject', 'text'),
+    ]
+
+
+class ContactsDatasetPipeline(BaseDatasetPipeline):
+    """Pipeline meta object for ContactsDataset."""
+
+    table_name = 'contacts_dataset'
+    source_url = '{0}/v4/dataset/contacts-dataset'.format(constants.DATAHUB_BASE_URL)
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('accepts_dit_email_marketing', 'accepts_dit_email_marketing', 'boolean'),
+        ('address_1', 'address_1', 'character varying(255)'),
+        ('address_2', 'address_2', 'character varying(255)'),
+        ('address_country__name', 'address_country', 'text'),
+        ('address_county', 'address_county', 'character varying(255)'),
+        ('address_postcode', 'address_postcode', 'character varying(255)'),
+        ('address_same_as_company', 'address_same_as_company', 'boolean'),
+        ('address_town', 'address_town', 'character varying(255)'),
+        ('company_id', 'company_id', 'uuid'),
+        ('created_on', 'date_added_to_datahub', 'date'),
+        ('email', 'email', 'character varying(255)'),
+        ('email_alternative', 'email_alternative', 'character varying(255)'),
+        ('id', 'id', 'uuid primary key'),
+        ('job_title', 'job_title', 'character varying(255)'),
+        ('name', 'contact_name', 'text'),
+        ('notes', 'notes', 'text'),
+        ('primary', 'is_primary', 'boolean'),
+        ('telephone_alternative', 'telephone_alternative', 'character varying(255)'),
+        ('telephone_number', 'phone', 'character varying(255)'),
+    ]
+
+
+class CompaniesDatasetPipeline(BaseDatasetPipeline):
+    """Pipeline meta object for CompaniesDataset."""
+
+    table_name = 'companies_dataset'
+    source_url = '{0}/v4/dataset/companies-dataset'.format(constants.DATAHUB_BASE_URL)
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('address_1', 'address_1', 'character varying(255)'),
+        ('address_2', 'address_2', 'character varying(255)'),
+        ('address_county', 'address_county', 'character varying(255)'),
+        ('address_country__name', 'address_country', 'character varying(255)'),
+        ('address_postcode', 'address_postcode', 'character varying(255)'),
+        ('address_town', 'address_town', 'character varying(255)'),
+        ('business_type__name', 'business_type', 'character varying(255)'),
+        ('company_number', 'company_number', 'character varying(255)'),
+        ('created_on', 'created_on', 'date'),
+        ('description', 'description', 'text'),
+        ('duns_number', 'duns_number', 'character varying(9)'),
+        (
+            'export_experience_category__name',
+            'export_experience',
+            'character varying(255)',
+        ),
+        ('id', 'id', 'uuid primary key'),
+        (
+            'is_number_of_employees_estimated',
+            'is_number_of_employees_estimated',
+            'boolean',
+        ),
+        ('is_turnover_estimated', 'is_turnover_estimated', 'boolean'),
+        ('name', 'name', 'character varying(255)'),
+        ('number_of_employees', 'number_of_employees', 'integer'),
+        ('one_list_tier__name', 'classification', 'character varying(255)'),
+        ('reference_code', 'cdms_reference_code', 'character varying(255)'),
+        ('registered_address_1', 'registered_address_1', 'character varying(255)'),
+        ('registered_address_2', 'registered_address_2', 'character varying(255)'),
+        (
+            'registered_address_country__name',
+            'registered_address_country',
+            'character varying(255)',
+        ),
+        (
+            'registered_address_county',
+            'registered_address_county',
+            'character varying(255)',
+        ),
+        (
+            'registered_address_postcode',
+            'registered_address_postcode',
+            'character varying(255)',
+        ),
+        (
+            'registered_address_town',
+            'registered_address_town',
+            'character varying(255)',
+        ),
+        ('sector_name', 'sector', 'character varying(255)'),
+        ('trading_names', 'trading_names', 'character varying(255)'),
+        ('turnover', 'turnover', 'bigint'),
+        ('uk_region__name', 'uk_region', 'character varying(255)'),
+        ('vat_number', 'vat_number', 'character varying(255)'),
+        ('website', 'website', 'character varying(255)'),
+    ]
+
+
+class AdvisersDatasetPipeline(BaseDatasetPipeline):
+    """Pipeline meta object for AdvisersDataset."""
+
+    table_name = 'advisers_dataset'
+    source_url = '{0}/v4/dataset/advisers-dataset'.format(constants.DATAHUB_BASE_URL)
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('id', 'id', 'uuid primary key'),
+        ('date_joined', 'date_joined', 'date'),
+        ('first_name', 'first_name', 'character varying(255)'),
+        ('last_name', 'last_name', 'character varying(255)'),
+        ('telephone_number', 'telephone_number', 'character varying(255)'),
+        ('contact_email', 'contact_email', 'character varying(255)'),
+        ('dit_team_id', 'team_id', 'uuid'),
+        ('is_active', 'is_active', 'boolean'),
+    ]
+
+
+class TeamsDatasetPipeline(BaseDatasetPipeline):
+    """Pipeline meta object for TeamsDataset."""
+
+    table_name = 'teams_dataset'
+    source_url = '{0}/v4/dataset/teams-dataset'.format(constants.DATAHUB_BASE_URL)
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('id', 'id', 'uuid primary key'),
+        ('name', 'name', 'character varying(255)'),
+        ('role__name', 'role', 'character varying(255)'),
+        ('uk_region__name', 'uk_region', 'character varying(255)'),
+        ('country__name', 'country', 'character varying(255)'),
+    ]
+
+
+class EventsDatasetPipeline(BaseDatasetPipeline):
+    """Pipeline meta object for EventsDataset."""
+
+    table_name = 'events_dataset'
+    source_url = '{0}/v4/dataset/events-dataset'.format(constants.DATAHUB_BASE_URL)
+    target_db = 'datasets_db'
+    start_date = datetime(2019, 11, 5)
+    end_date = None
+    schedule_interval = '@daily'
+    field_mapping = [
+        ('address_1', 'address_1', 'character varying(255)'),
+        ('address_2', 'address_2', 'character varying(255)'),
+        ('address_country__name', 'address_country', 'character varying(255)'),
+        ('address_county', 'address_county', 'character varying(255)'),
+        ('address_postcode', 'address_postcode', 'character varying(255)'),
+        ('address_town', 'address_town', 'character varying(255)'),
+        ('created_on', 'created_on', 'timestamp with time zone'),
+        ('end_date', 'end_date', 'date'),
+        ('event_type__name', 'event_type', 'character varying(255)'),
+        ('id', 'id', 'uuid primary key'),
+        ('lead_team_id', 'lead_team_id', 'uuid'),
+        ('location_type__name', 'location_type', 'character varying(255)'),
+        ('name', 'name', 'character varying(255)'),
+        ('notes', 'notes', 'text'),
+        ('organiser_id', 'organiser_id', 'uuid'),
+        ('service_name', 'service_name', 'character varying(255)'),
+        ('start_date', 'start_date', 'date'),
+        ('team_ids', 'team_ids', 'text []'),
+        ('uk_region__name', 'uk_region', 'character varying(255)'),
+    ]
+
+
+for pipeline in BaseDatasetPipeline.__subclasses__():
+    globals()[pipeline.__name__ + '__dag'] = pipeline.get_dag()
