+++ dataflow/utils.py
@@ -4,15 +4,6 @@ from airflow.hooks.postgres_hook import PostgresHook
 from airflow.operators.postgres_operator import PostgresOperator
 
 
-def get_defined_pipeline_classes_by_key(module, key):
-    """Return all class objects in given module which contains given key in its name."""
-    return [
-        cls
-        for name, cls in module.__dict__.items()
-        if isinstance(cls, type) and key in name
-    ]
-
-
 class XCOMIntegratedPostgresOperator(PostgresOperator):
     """Custom PostgresOperator which push query result into XCOM."""
